revert unnecessary codegen cli changes
